requirements.txt
Flask==2.3.3
python-dotenv==1.1.1
python-telegram-bot==20.7
Werkzeug==3.1.3
psycopg2-binary==2.9.10